% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_performance-assessment.R
\name{create_performance_review}
\alias{create_performance_review}
\title{Create a responsematch 'Shiny' Survey Performance Review}
\usage{
create_performance_review(name, path)
}
\arguments{
\item{name}{The name of the survey application to create}

\item{path}{The path specifying where the survey application should be
created. Default is the current working directory.}
}
\value{
A new RStudio Project the infrastructure needed to launch a new
performance review in Shiny.
}
\description{
This function can be used to create a Shiny Survey application. It takes in
two parameters: \code{name}, which specifies the name of the performance
review and \code{path}, which specifies where the application should be
created.
}
\details{
Assuming the data should be stored to Google Drive/Google Sheets (see
function arguments for more details), running this function will result in
your internet browser will open and you will be asked to authenticate your
Google Drive and Google Sheets information. This is necessary to remotely
store respondent's data.

After Google Drive and Google Sheets authentication, a new \url{RStudio
Project}{https://r4ds.had.co.nz/workflow-projects.html} will be created at
the specified path, with a suffix 'survey-project'.

The RStudio project will open in a new R Session automatically. In the file
pane, there will be a folder with a suffix '-app', containing an 'app.R' file
-- defining the user-interface and server-side logic of the survey -- a 'www'
folder -- housing some example questions, RMarkdown reports, and custom
JavaScript -- and an 'R' directory containing additional functions used by
the application.

The Shiny application is fully functioning at this point, however, you must
modify the survey questions and parameterized RMarkdown report located in the
'www' subdirectory.

In order to deploy the survey to \href{https://shinyapps.io/}{shinyapps.io}, you
may click the 'deploy' button at the top right of the RStudio IDE. For more
details on deployment, please see my \url{blog post on getting started with
[shinyapps.io]}{https://www.jdtrat.com/blog/getting-started-shinyapps/}.
}
\examples{

if (interactive()) {
create_performance_review(name = "testPerformanceReview")
}

}
