% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_survey.R
\name{create_survey}
\alias{create_survey}
\title{Create a responsematch 'Shiny' Survey Application}
\usage{
create_survey(survey_name, path = getwd())
}
\arguments{
\item{survey_name}{The name of the survey application to create, e.g. "GLI".}

\item{path}{The path specifying where the survey application should be
created. Default is the current working directory.}
}
\value{
A Shiny application inside a folder of \code{survey_name}.
It contains an 'app.R' file -- defining the user-interface and server-side
logic of the survey -- a 'www' folder -- housing the survey questions,
RMarkdown report, and custom JavaScript -- and an 'R' directory containing
additional functions used by the application.
}
\description{
This function can be used to create a Shiny Survey application. It takes in
two parameters: \code{survey_name}, which specifies the name of the survey --
e.g. "GLI" -- and \code{path}, which specifies where the application should
be created.
}
\details{
When run, your internet browser will open and you will be asked to
authenticate your Google Drive and Google Sheets information. This is
necessary to remotely store respondent's data. The resulting Shiny app will
be created in a folder specified by the \code{path} argument. Before publishing
the survey, you must modify the survey questions and parameterized RMarkdown
report  located in the 'www' subdirectory.
}
\examples{

if (interactive()) {
create_survey(survey_name = "testSurvey")
}

}
