% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-passwords.R
\name{gen_pw_from_char}
\alias{gen_pw_from_char}
\title{Generate Password from Character String}
\usage{
gen_pw_from_char(char, key, pw_length = 10)
}
\arguments{
\item{char}{A character string from which to generate a password.}

\item{key}{A character string representing a secret key for generating the
password with the given character string.}

\item{pw_length}{(numeric) Length of the password}
}
\value{
A character string of length \code{pw_length} generated with a seed
derived from the \code{char} provided (a public facing string, such as a name)
as wekk as the \code{key} (a secret string known to the password generator).
}
\description{
Generate a password from a non-secret string \code{char} with a
secret \code{key}.
}
\examples{

# Create a password for the name "Jonathan" using the secret key "password".
gen_pw_from_char(char = "Jonathan", key = "password", pw_length = 10)

}
