% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_setup.R
\name{setup_survey_sheet}
\alias{setup_survey_sheet}
\title{Setup survey infrastructure in Google Drive and Google Sheets}
\usage{
setup_survey_sheet(survey_name)
}
\arguments{
\item{survey_name}{The name of the new survey to create}
}
\value{
NA; used to establish survey infrastructure with Google Drive and Sheets.
}
\description{
This function takes in a survey name, e.g. "GLI", and will create a folder in
Google Drive with a corresponding Google Sheet for all survey data to be
written.
}
\examples{

if (interactive()) {
setup_survey_sheet("testSurvey")
}

}
