% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_assessment_setup.R
\name{setup_performance_review}
\alias{setup_performance_review}
\title{Setup survey infrastructure in Google Drive and Google Sheets}
\usage{
setup_performance_review(name)
}
\arguments{
\item{name}{The name of the new performance assessment}
}
\value{
A named list of Google Sheets IDs for customizing the Performance Review Shiny app \code{c( 'employee_response_folder_id', 'employe_survey_metadata_id', 'employee_login_data_id', 'employe_survey_responses_id', 'employee_dependencies_id' )}. This named list is used internally \code{\link[=create_performance_review]{create_performance_review()}}.
}
\description{
This function takes in a name and will create a folder in
Google Drive with a corresponding Google Sheet for all survey data to be
written.
}
\examples{

if (interactive()) {
setup_performance_review("testPA")
}

}
